test: add device_long unit tests
